Move Distributor interface back to runtime.
Distributor is an interface consumed by the runtime package.
